+++ src/BusinessLogic/BootstrapComponent.php
@@ -12,14 +12,14 @@ use Packlink\BusinessLogic\Controllers\DTO\DashboardStatus;
 use Packlink\BusinessLogic\Controllers\ShippingMethodController;
 use Packlink\BusinessLogic\Country\Country;
 use Packlink\BusinessLogic\Country\CountryService;
-use Packlink\BusinessLogic\Country\RegistrationCountry;
 use Packlink\BusinessLogic\Country\WarehouseCountryService;
 use Packlink\BusinessLogic\DTO\FrontDtoFactory;
 use Packlink\BusinessLogic\DTO\ValidationError;
+use Packlink\BusinessLogic\FileResolver\FileResolverService;
 use Packlink\BusinessLogic\Http\DTO\ParcelInfo;
 use Packlink\BusinessLogic\Http\Proxy;
-use Packlink\BusinessLogic\Language\Interfaces\TranslationService as TranslationServiceInterface;
-use Packlink\BusinessLogic\Language\TranslationService;
+use Packlink\BusinessLogic\CountryLabels\Interfaces\CountryService as LabelServiceInterface;
+use Packlink\BusinessLogic\CountryLabels\CountryService as CountryLabelService;
 use Packlink\BusinessLogic\Location\LocationService;
 use Packlink\BusinessLogic\Order\OrderService;
 use Packlink\BusinessLogic\OrderShipmentDetails\OrderShipmentDetailsService;
@@ -171,9 +171,24 @@ class BootstrapComponent extends \Logeecom\Infrastructure\BootstrapComponent
         );
 
         ServiceRegister::registerService(
-            TranslationServiceInterface::CLASS_NAME,
+            FileResolverService::CLASS_NAME,
             function () {
-                return new TranslationService();
+                return new FileResolverService(
+                    array(
+                        __DIR__ . '/../Brands/Packlink/Resources/countries',
+                        __DIR__ . '/Resources/countries',
+                    )
+                );
+            }
+        );
+
+        ServiceRegister::registerService(
+            LabelServiceInterface::CLASS_NAME,
+            function () {
+                /** @var FileResolverService $fileResolverService */
+                $fileResolverService = ServiceRegister::getService(FileResolverService::CLASS_NAME);
+
+                return new CountryLabelService($fileResolverService);
             }
         );
 
@@ -217,7 +232,6 @@ class BootstrapComponent extends \Logeecom\Infrastructure\BootstrapComponent
         FrontDtoFactory::register(ParcelInfo::CLASS_KEY, ParcelInfo::CLASS_NAME);
         FrontDtoFactory::register(DashboardStatus::CLASS_KEY, DashboardStatus::CLASS_NAME);
         FrontDtoFactory::register(Country::CLASS_KEY, Country::CLASS_NAME);
-        FrontDtoFactory::register(RegistrationCountry::CLASS_KEY, RegistrationCountry::CLASS_NAME);
         FrontDtoFactory::register(RegistrationRequest::CLASS_KEY, RegistrationRequest::CLASS_NAME);
         FrontDtoFactory::register(RegistrationLegalPolicy::CLASS_KEY, RegistrationLegalPolicy::CLASS_NAME);
         FrontDtoFactory::register(ShippingPricePolicy::CLASS_KEY, ShippingPricePolicy::CLASS_NAME);