+++ src/BusinessLogic/Brand/Exceptions/PlatformCountryNotSupportedByBrandException.php
@@ -1,0 +1,16 @@
+<?php
+
+
+namespace Packlink\BusinessLogic\Brand\Exceptions;
+
+use Logeecom\Infrastructure\Exceptions\BaseException;
+
+/**
+ * Class PlatformCountryNotSupportedByBrandException
+ *
+ * @package Packlink\BusinessLogic\Brand\Exceptions
+ */
+class PlatformCountryNotSupportedByBrandException extends BaseException
+{
+
+}