+++ src/BusinessLogic/Controllers/RegistrationController.php
@@ -4,7 +4,9 @@ namespace Packlink\BusinessLogic\Controllers;
 
 use Logeecom\Infrastructure\Configuration\Configuration;
 use Logeecom\Infrastructure\ServiceRegister;
+use Packlink\BusinessLogic\Brand\BrandConfigurationService;
 use Packlink\BusinessLogic\DTO\FrontDtoFactory;
+use Packlink\BusinessLogic\CountryLabels\Interfaces\CountryService;
 use Packlink\BusinessLogic\Registration\RegistrationInfoService;
 use Packlink\BusinessLogic\Registration\RegistrationRequest;
 use Packlink\BusinessLogic\Registration\RegistrationService;
@@ -21,45 +23,13 @@ class RegistrationController
      */
     protected $configService;
     /**
-     * List of terms and conditions URLs for different country codes.
-     *
-     * @var array
+     * @var BrandConfigurationService
      */
-    private static $termsAndConditionsUrls = array(
-        'EN' => 'https://support-pro.packlink.com/hc/en-gb/articles/360010011480',
-        'ES' => 'https://pro.packlink.es/terminos-y-condiciones/',
-        'DE' => 'https://pro.packlink.de/agb/',
-        'FR' => 'https://pro.packlink.fr/conditions-generales/',
-        'IT' => 'https://pro.packlink.it/termini-condizioni/',
-        'AT' => 'https://support-pro.packlink.com/hc/de/articles/360010011480',
-        'NL' => 'https://support-pro.packlink.com/hc/nl/articles/360010011480',
-        'BE' => 'https://support-pro.packlink.com/hc/nl/articles/360010011480',
-        'PT' => 'https://support-pro.packlink.com/hc/pt/articles/360010011480',
-        'TR' => 'https://support-pro.packlink.com/hc/tr/articles/360010011480',
-        'IE' => 'https://support-pro.packlink.com/hc/en-gb/articles/360010011480',
-        'GB' => 'https://support-pro.packlink.com/hc/en-gb/articles/360010011480',
-        'HU' => 'https://support-pro.packlink.com/hc/hu/articles/360010011480',
-    );
+    protected $brandConfigurationService;
     /**
-     * List of terms and conditions URLs for different country codes.
-     *
-     * @var array
+     * @var CountryService
      */
-    private static $privacyPolicyUrls = array(
-        'EN' => 'https://support-pro.packlink.com/hc/en-gb/articles/360010011560',
-        'ES' => 'https://support-pro.packlink.com/hc/es-es/articles/360010011560-Pol%C3%ADtica-de-Privacidad',
-        'DE' => 'https://support-pro.packlink.com/hc/de/articles/360010011560-Datenschutzerkl%C3%A4rung-der-Packlink-Shipping-S-L-',
-        'FR' => 'https://support-pro.packlink.com/hc/fr-fr/articles/360010011560-Politique-de-confidentialit%C3%A9',
-        'IT' => 'https://support-pro.packlink.com/hc/it/articles/360010011560-Politica-di-Privacy',
-        'AT' => 'https://support-pro.packlink.com/hc/de/articles/360010011480',
-        'NL' => 'https://support-pro.packlink.com/hc/nl/articles/360010011560',
-        'BE' => 'https://support-pro.packlink.com/hc/nl/articles/360010011560',
-        'PT' => 'https://support-pro.packlink.com/hc/pt/articles/360010011560',
-        'TR' => 'https://support-pro.packlink.com/hc/tr/articles/360010011560',
-        'IE' => 'https://support-pro.packlink.com/hc/en-gb/articles/360010011560',
-        'GB' => 'https://support-pro.packlink.com/hc/en-gb/articles/360010011560',
-        'HU' => 'https://support-pro.packlink.com/hc/hu/articles/360010011560',
-    );
+    protected $countryService;
 
     /**
      * Gets the data needed for a registration page.
@@ -74,15 +44,19 @@ class RegistrationController
         $registrationInfoService = ServiceRegister::getService(RegistrationInfoService::CLASS_NAME);
         $registrationData = $registrationInfoService->getRegistrationInfoData();
 
+        $brand = $this->getBrandConfigurationService()->get();
+
         return array(
             'context' => $this->getConfigService()->getContext(),
             'email' => $registrationData->getEmail(),
             'phone' => $registrationData->getPhone(),
             'source' => $registrationData->getSource(),
-            'termsAndConditionsUrl' => !empty(self::$termsAndConditionsUrls[$country]) ?
-                self::$termsAndConditionsUrls[$country] : self::$termsAndConditionsUrls[self::DEFAULT_COUNTRY],
-            'privacyPolicyUrl' => !empty(self::$privacyPolicyUrls[$country]) ?
-                self::$privacyPolicyUrls[$country] : self::$privacyPolicyUrls[self::DEFAULT_COUNTRY],
+            'termsAndConditionsUrl' => $this->getCountryService()
+                ->getLabel(strtolower($country), 'register.termsAndConditionsUrl', static::DEFAULT_COUNTRY),
+            'privacyPolicyUrl' => $this->getCountryService()
+                ->getLabel(strtolower($country), 'register.privacyPolicyUrl', static::DEFAULT_COUNTRY),
+            'platform_country' => in_array($country, $brand->platformCountries, true) ?
+                $country : $brand->platformCountries[0],
         );
     }
 
@@ -98,11 +72,14 @@ class RegistrationController
      * @throws \Packlink\BusinessLogic\DTO\Exceptions\FrontDtoNotRegisteredException
      * @throws \Packlink\BusinessLogic\DTO\Exceptions\FrontDtoValidationException
      * @throws \Packlink\BusinessLogic\Registration\Exceptions\UnableToRegisterAccountException
+     * @throws \Packlink\BusinessLogic\Brand\Exceptions\PlatformCountryNotSupportedByBrandException
      */
     public function register(array $payload)
     {
-        $payload['platform'] = 'PRO';
-        $payload['language'] = $this->getLanguage($payload['platform_country']);
+        $brand = $this->getBrandConfigurationService()->get();
+
+        $payload['platform'] = $brand->platformCode;
+        $payload['language'] = $this->getLanguage();
 
         if (isset($payload['source'])) {
             $payload['source'] = 'https://' . str_replace(array('http://', 'https://'), '', $payload['source']);
@@ -153,7 +130,35 @@ class RegistrationController
     }
 
     /**
-     * Returns shop language in format which Packlink expects, based on user's platform country.
+     * Returns an instance of brand configuration service.
+     *
+     * @return BrandConfigurationService
+     */
+    protected function getBrandConfigurationService()
+    {
+        if ($this->brandConfigurationService === null) {
+            $this->brandConfigurationService = ServiceRegister::getService(BrandConfigurationService::CLASS_NAME);
+        }
+
+        return $this->brandConfigurationService;
+    }
+
+    /**
+     * Returns an instance of country service.
+     *
+     * @return CountryService
+     */
+    protected function getCountryService()
+    {
+        if ($this->countryService === null) {
+            $this->countryService = ServiceRegister::getService(CountryService::CLASS_NAME);
+        }
+
+        return $this->countryService;
+    }
+
+    /**
+     * Returns shop language in format which Packlink expects.
      *
      * @param string $platformCountry
      *
@@ -168,6 +173,7 @@ class RegistrationController
             'IT' => 'it_IT',
         );
 
+        $locale = Configuration::getUICountryCode();
         $language = 'en_GB';
 
         if (array_key_exists($platformCountry, $supportedLanguages)) {