+++ src/BusinessLogic/Scheduler/Models/WeeklySchedule.php
@@ -34,7 +34,18 @@ class WeeklySchedule extends Schedule
      *
      * @var array
      */
-    protected $fields = array('id', 'queueName', 'minute', 'hour', 'day', 'month', 'lastWeek', 'weeks', 'context');
+    protected $fields = array(
+        'id',
+        'queueName',
+        'minute',
+        'hour',
+        'day',
+        'month',
+        'lastWeek',
+        'weeks',
+        'context',
+        'brand',
+    );
     /**
      * Last week flag.
      *