+++ src/DemoUI/run.sh
@@ -2,6 +2,10 @@
 
 composer install
 
-xdg-open http://localhost:7000/Views/index.php
+PL_PLATFORM="${1:-"PRO"}"
+
+export PL_PLATFORM
+
+xdg-open http://localhost:7000/Views/$PL_PLATFORM/index.php
 
 cd $PWD/src && php -S localhost:7000