+++ tests/BusinessLogic/Country/CountryServiceTest.php
@@ -2,6 +2,7 @@
 
 namespace Logeecom\Tests\BusinessLogic\Country;
 
+use Logeecom\Infrastructure\Configuration\Configuration;
 use Logeecom\Infrastructure\ServiceRegister;
 use Logeecom\Tests\BusinessLogic\Common\BaseTestWithServices;
 use Packlink\BusinessLogic\Country\CountryService;
@@ -15,6 +16,7 @@ class CountryServiceTest extends BaseTestWithServices
 {
     public function testGetSupportedCountries()
     {
+        Configuration::setUICountryCode('en');
         /** @var CountryService $service */
         $service = ServiceRegister::getService(CountryService::CLASS_NAME);
 
@@ -26,9 +28,5 @@ class CountryServiceTest extends BaseTestWithServices
         $this->assertEquals('Spain', $countries['ES']->name);
         $this->assertEquals('ES', $countries['ES']->code);
         $this->assertEquals('28001', $countries['ES']->postalCode);
-        $this->assertEquals(
-            'https://auth.packlink.com/es-ES/test-system/registro?platform=PRO&platform_country=ES',
-            $countries['ES']->registrationLink
-        );
     }
 }